+++ source/extensions/omni.flux.stage_manager.plugin.interaction.usd/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.5.3]
+### Added
+- Added auto scroll to selection functionality
+
 ## [1.5.2]
 ### Fixed
 - Don't select all children when selecting an item